EE-208: Move partial ordering test of AccessRights to tests crate.
