+++ execution-engine/common/src/lib.rs
@@ -8,13 +8,6 @@
     maybe_uninit
 )]
 
-#[cfg(test)]
-#[macro_use]
-extern crate proptest;
-
-#[cfg(test)]
-extern crate gens;
-
 extern crate alloc;
 #[macro_use]
 extern crate uint;
